refactor memory utils
